Downloads/log-aggregator/Application/requirements.txt
# Log Aggregator Dependencies

# Core Framework
Flask==2.3.3
Werkzeug==2.3.7

# Environment Configuration
python-dotenv==1.0.1

# Optional: PDF Report Generation
reportlab==4.2.5

# Optional: AWS Bedrock Integration
boto3
botocore

# Development (optional)
Pillow==10.0.0  # Image support for reportlab (optional)
